emmm...

+++ base/src/main/java/com/orc/sandbox/base/dao/UserDao.java
@@ -1,6 +1,7 @@
 package com.orc.sandbox.base.dao;
 
 import com.orc.sandbox.base.entity.User;
+import com.orc.sandbox.common.dao.BaseDao;
 
-public interface UserDao extends BaseDao<User, Integer> {
+public interface UserDao extends BaseDao<User, Long> {
 }
